+++ lib/res/model/show_dialog_parameters.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+// Use this class in changeable dialog
+class ShowDialogParameters {
+  final BuildContext context;
+
+  final bool barrierDismissible;
+
+  final Widget dialogChildWidget;
+
+  final Function(dynamic)? actionOnCloseModal;
+
+  ShowDialogParameters({
+    this.actionOnCloseModal,
+    this.barrierDismissible = true,
+    required this.context,
+    required this.dialogChildWidget,
+  });
+}